Refactor PostgreSQLAuthenticationException
